First commit
I felt like making a project that could be done in a short amount of time. I'm surprised at just how little time was spent building this project; roughly half an hour. Therefore, I'm going to spend a little more time on the design. I also want to either make the logic more complex or simplify parts of the Javascript, or both.