Final commit
I had thought of trying to add more functionality to this such as features that allow the user to select the amount of characters they want their password to be. I would also have liked to include a form feature that requests the user's name and then manipulates this information to construct a unique password. 

Unfortunately, I have been ill for the past 8 days and don't have the strength to do much today. Therefore, seeing that every Tom, Dick and Harry makes a random generator, I have settled with simply making the design look more presentable and responsive.